TD-22 change folder structure amnd file names
