+++ .env
@@ -2,3 +2,4 @@ CLIENT_ID=792733349077-6hmut9k86hiecsb4lgpm0ehoiloiir93.apps.googleusercontent.c
 CLIENT_SECRET=IzBORHvI4MIvNK3o5oFJ8IYl
 CALLBACK_URL=http://localhost:3000/user/auth/google/callback
 PORT=3000
+RND_STRING=abcdefghijklm0123456789nopqrstuvwxyz0123456789ABCDEFGHIJKLM0123456789NOPQRSTUVWXYZ0123456789
