+++ config/language_message.js
@@ -1,0 +1,13 @@
+module.exports = {
+    validate:'Please enter all the fields',
+    status_true:'ok',
+    status_false:'not ok',
+    register_validate: 'User already registered!',
+    empty_validate: 'Please enter all the fields',
+    success_login: 'User successfully login',
+    incorrect_password: 'Incorrect Password!',
+    incorrect_email: 'Incorrect Email!',
+    user_edit: 'User edited successfully!',
+    not_found: 'No user found with this id!',
+    data_delete: 'Data successfully deleted!'
+};
