+++ package.json
@@ -31,6 +31,7 @@
     "generate-password": "^1.4.0",
     "jest": "^23.4.2",
     "jsonwebtoken": "^8.3.0",
+    "jwt-simple": "^0.5.1",
     "knex": "^0.15.2",
     "nodemailer": "^4.6.7",
     "nodemon": "^1.18.3",
